# apps/daemon/openapi.json
{
  "components": {
    "schemas": {
...
      "Attachment": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "kind": {
            "enum": [
              "audio",
              "image",
              "pdf"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pages": {
            "default": 1,
            "title": "Pages",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "name",
          "kind"
        ],
        "title": "Attachment",
        "type": "object"
      },
      "AttachmentInfo": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "generated": {
            "default": false,
            "title": "Generated",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "audio",
              "image",
              "pdf",
              "text",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "modified": {
            "title": "Modified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "referenced": {
            "title": "Referenced",
            "type": "boolean"
          },
          "referenced_by": {
            "items": {
              "type": "string"
            },
            "title": "Referenced By",
            "type": "array"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "name",
          "kind",
          "size",
          "modified",
          "referenced",
          "referenced_by"
        ],
        "title": "AttachmentInfo",
        "type": "object"
      },
...
      "ExtractRequest": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          },
          "output": {
            "default": "toggle",
            "enum": [
              "toggle",
              "page"
            ],
            "title": "Output",
            "type": "string"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id",
          "file",
          "name"
        ],
        "title": "ExtractRequest",
        "type": "object"
      },
      "Extraction": {
        "properties": {
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "file": {
            "default": "",
            "title": "File",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageDoc"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "default": "Waiting for the local model\u2026",
            "title": "Progress",
            "type": "string"
          },
          "status": {
            "default": "running",
            "enum": [
              "running",
              "done",
              "error",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "Extraction",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MoveMediaRequest": {
        "properties": {
          "base_hash": {
            "title": "Base Hash",
            "type": "string"
          },
          "block": {
            "maxLength": 10000,
            "title": "Block",
            "type": "string"
          },
          "file": {
            "title": "File",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "target_path": {
            "title": "Target Path",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "target_path",
          "file",
          "block",
          "base_hash"
        ],
        "title": "MoveMediaRequest",
        "type": "object"
      },
...
      "TrashedAttachment": {
        "properties": {
          "trash_id": {
            "title": "Trash Id",
            "type": "string"
          }
        },
        "required": [
          "trash_id"
        ],
        "title": "TrashedAttachment",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "graite-daemon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/media/attach-recording": {
      "post": {
        "operationId": "attach_recording_api_v1_media_attach_recording_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Attach Recording Api V1 Media Attach Recording Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Recording",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/attachments": {
      "delete": {
        "description": "Move one attachment to the trash. Pages are never rewritten.",
        "operationId": "trash_attachment_api_v1_media_attachments_delete",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrashedAttachment"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Still used on a page; repeat with force=true."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trash Attachment",
        "tags": [
          "media"
        ]
      },
      "get": {
        "description": "Every file in this page's _assets folder, and which pages still use it.",
        "operationId": "list_attachments_api_v1_media_attachments_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentInfo"
                  },
                  "title": "Response List Attachments Api V1 Media Attachments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Attachments",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/extract": {
      "post": {
        "operationId": "extract_api_v1_media_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extraction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/file": {
      "get": {
        "operationId": "get_file_api_v1_media_file_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/info": {
      "get": {
        "operationId": "info_api_v1_media_info_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Info Api V1 Media Info Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Info",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/jobs/{job_id}": {
      "delete": {
        "operationId": "cancel_job_api_v1_media_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extraction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "media"
        ]
      },
      "get": {
        "operationId": "get_job_api_v1_media_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extraction"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/location": {
      "get": {
        "operationId": "location_api_v1_media_location_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Location Api V1 Media Location Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/move": {
      "post": {
        "operationId": "move_media_api_v1_media_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Move Media Api V1 Media Move Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/preview": {
      "get": {
        "operationId": "preview_api_v1_media_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/upload": {
      "post": {
        "operationId": "upload_api_v1_media_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "media"
        ]
      }
    },
...
  }
}